Task 1 complete
In this task, I defined and implemented a function that yields the next Fibonacci number in the sequence each time its called into the program. I first added the function signature and then created a stub implementation. After this, I implemented the Fibonacci number yield function and returned either the value 1 or the next value in the sequence depending on the boolean value of the reset flag. All tests are passing for this task.