Task 2 complete
In this task, I created and implemented two new functions (swap() and threeSort()) that take char types as inputs to be swapped and sorted. I reused the the two previous functions (swap() and threeSort()) that took in int types as inputs, and changed the type to that of type char. I also updated the function documentation to reflect this. All tests are passing for this task.